Create data shuffling method as a pre-step to pretokenization
Since the exported data from wikipeadi is ordered by search term, we add
a method that load in memory all text data, shuffle all the articles and
save them back into shards.

This will faciliate the role of the data iterator to ensure that workers get
attributed random articles and not articles from the same search.