Adapt transformer model from llama (facebook) and llama.c (Karpathy)
We copy the llama code from facebook research and in particular the
adaptation made by Andrej Karpathy in his llama.c repo.

The version is slightly simplified: I do not use of the kv
cache and the multi-group query since we plan to train small
language model (we might add them later on just to see the effect
on inference performance).

In the transformer block, I don't rely on the module.forward()
method but instead directly call the module() for the forward
pass. I believe (?) calling the .forward might result in unexpected
behavior such as the dropout not working as expected at inference
but I might be wrong.