Compute approximate tokens vs. params optimal values
I fork the notebook 'scaling_laws' from A. Karpathy repo nano-gpt.
This is essentially the same reproduction as the Chinchilla paper
except the params count of this model is slightly different from
nano-gpt.

Also I add a crude simulation on how to maximize the number of
parameters for a given FLOPS budget. This should be refined to
get more consistent model architecture solutions.